v2.32: UI evolution, deep debt, coverage push — 4,033 tests, 75% coverage
- TUI deprecated and feature-gated behind `deprecated-tui`; petalTongue is canonical UI
- Logic/IO separation across ~20 CLI and mode files (pure functions extracted)
- Capability discovery evolved: taxonomy-based, deprecated name lookups
- Credential security: HMAC-SHA256 integrity, versioned format, 0o600 permissions
- live_service.rs placeholder replaced with real /proc/uptime + SystemInspector
- 63+ new tests across 16 modules; 885 new tests total
- Hardcoded primal names eliminated from production discovery paths
- nix-to-rustix migration assessment created
- Dead code audit: wire format annotated, future-use documented
- Python examples and stale debris cleaned
- Root docs, CHANGELOG, wateringHole handoff updated
- TODO audit: 4 valid wire-up TODOs remain, 0 outdated

Made-with: Cursor

+++ crates/biomeos-spore/src/usb.rs
@@ -1,3 +1,6 @@
+// SPDX-License-Identifier: AGPL-3.0-only
+// Copyright 2025 ecoPrimals Project
+
 //! USB device discovery and management
 //!
 //! Provides capability-based USB device discovery without hardcoding.
@@ -83,8 +86,8 @@ async fn probe_device(path: PathBuf) -> SporeResult<UsbDevice> {
             .await
         {
             Ok(Ok(st)) => {
-                let avail = st.blocks_available() as u64 * st.fragment_size() as u64;
-                let total = st.blocks() as u64 * st.fragment_size() as u64;
+                let avail = st.blocks_available() * st.fragment_size();
+                let total = st.blocks() * st.fragment_size();
                 (avail, total)
             }
             _ => {
